requirements.txt
langgraph>=0.4.0
langchain>=0.3.20
langchain-google-genai>=4.2.2
python-dotenv==1.0.1
